+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/weapons/armour_custom/marines2/config.hpp
@@ -1,1 +1,1 @@
-#include "group2\config.hpp"
+#include "group1\config.hpp"
